[ADD] updated code

+++ logistic/custom_logistic/models.py
@@ -1,11 +1,14 @@
 from odoo import models, fields, api
-from datetime import datetime,date
+from datetime import datetime,date,timedelta
+import xlsxwriter
+import os
 
 class Exportlogic(models.Model):
 	_name = 'export.logic'
 	_rec_name = 'sr_no'
 
 	customer         = fields.Many2one('res.partner',string="Customer",required=True)
+	by_customer      = fields.Many2one('by.customer', string="By Customer")
 	sr_no       	 = fields.Char(string="SR No", readonly=True)
 	bill_bol         = fields.Boolean(string="B/L")
 	cont_bol         = fields.Boolean(string="Cont")
@@ -13,12 +16,15 @@ class Exportlogic(models.Model):
 	bill_types       = fields.Char(string="Billing Type")
 	our_job_no       = fields.Char(string="Our Job No", readonly=True,)
 	customer_ref     = fields.Char(string="Customer Ref")
+	cust_ref_inv     = fields.Char(string="Customer Ref Inv No")
 	shipper_date     = fields.Date(string="DOC Received Date",default=date.today())
 	mani_date        = fields.Date(string="Manifest Received Date")
 	date             = fields.Date(string="Date",required=True,default=date.today())
 	bill_no          = fields.Char(string="B/L Number")
+	rot_no           = fields.Char(string="Rotation Number/Sequence Number")
 	bill_attach      = fields.Binary(string=" ")
-	eta              = fields.Char(string="ETA")
+	eta              = fields.Date(string="ETA")
+	etd              = fields.Date(string="ETD")
 	about            = fields.Char(string="On Or About")
 	twen_ft          = fields.Integer(string="20 ft")
 	fort_ft          = fields.Integer(string="40 ft")
@@ -29,14 +35,17 @@ class Exportlogic(models.Model):
 	pre_bayan        = fields.Date(string="Pre Bayan Date")
 	custom_exam      = fields.Boolean(string="Open Custom Examination")
 	bayan_date       = fields.Date(string="Initial Bayan Date")
-	shutt_start_date = fields.Date(string="Shuttling Start Date")
+	shutl_start_date = fields.Date(string="Shuttling Start Date")
 	fin_bayan_date   = fields.Date(string="Final Bayan Date")
-	shutt_end_date   = fields.Date(string="Shuttling End Date")
+	shutl_end_date   = fields.Date(string="Shuttling End Date")
 	status           = fields.Many2one('import.status',string="Status")
 	fri_id           = fields.Many2one('freight.forward', string="Freight Link")
 	site             = fields.Many2one('import.site',string="Site")
+	acc_link         = fields.Many2one('account.invoice',string="Invoice",readonly=True)
 	remarks          = fields.Text(string="Remarks")
 	vessel_date 	 = fields.Date(string="Vessel Arrival Date")
+	vessel_name      = fields.Char(string="Vessel Name")
+	s_supplier       = fields.Many2one('res.partner',string="Shipping Line")
 	export_link 	 = fields.One2many('logistic.export.tree','export_tree')
 	export_id 	     = fields.One2many('export.tree','crt_tree')
 	export_serv 	 = fields.One2many('logistic.service.tree','service_tree')
@@ -49,7 +58,11 @@ class Exportlogic(models.Model):
 					 ('custom_exam', 'Custom Examination'),
 					 ('done', 'Done'),
 					 ],default='draft')
+	_sql_constraints = [
+	('customer_ref', 'unique(customer_ref)','This customer reference already esixts!')
+	]
 	
+	tick  = fields.Boolean()
 	
 
 	@api.onchange('custom_exam')
@@ -73,52 +86,50 @@ class Exportlogic(models.Model):
 			self.bill_types = records.bill_type
 
 
-	@api.onchange('customer','bill_types')
+	@api.onchange('customer','by_customer')
 	def get_tree_value(self):
 		if self.customer:
 			if self.bill_types == "B/L Number":
 				records = self.env['res.partner'].search([('id','=',self.customer.id)])
+				for x in records.bl_id:
+					if self.by_customer == x.by_customer:
 
-				export_serv = []
-				delete = []
-				delete = delete.append(2)
-				self.export_serv = delete
+						export_serv = []
+						delete = []
+						delete = delete.append(2)
+						self.export_serv = delete
 
-				inv = []
-				for invo in records.bl_id:
-					inv.append({
-						'sevr_charge':invo.charges_serv,
-						'sevr_type':invo.charges_type.id,
-						'service_tree':self.id,
-						})
-				
-				self.export_serv = inv
-				inv=[]
+						inv = []
+						for invo in x:
+							inv.append({
+								'sevr_charge':invo.charges_serv,
+								'sevr_type':invo.charges_type.id,
+								'service_tree':self.id,
+								})
+						
+						self.export_serv = inv
+						inv=[]
 
 			if self.bill_types == "Container Wise":
 				records = self.env['res.partner'].search([('id','=',self.customer.id)])
+				for x in records.cont_id:
+					if self.by_customer == x.by_customer:
+						cont_serv = []
+						delete = []
+						delete = delete.append(2)
+						self.cont_serv = delete
 
-				cont_serv = []
-				delete = []
-				delete = delete.append(2)
-				self.cont_serv = delete
-
-				inv = []
-				for invo in records.cont_id:
-					inv.append({
-						'sevr_charge_cont':invo.charges_serv,
-						'sevr_type_cont':invo.charges_type.id,
-						'type_contt':invo.cont_type,
-						'cont_serv':self.id,
-						})
-				
-				self.cont_serv = inv
-				inv=[]
-
-
-
-
-
+						inv = []
+						for invo in x:
+							inv.append({
+								'sevr_charge_cont':invo.charges_serv,
+								'sevr_type_cont':invo.charges_type.id,
+								'type_contt':invo.cont_type,
+								'cont_serv':self.id,
+								})
+						
+						self.cont_serv = inv
+						inv=[]
 
 	@api.onchange('bill_types')
 	def get_tree(self):
@@ -140,7 +151,6 @@ class Exportlogic(models.Model):
 	def create(self, vals):
 		vals['sr_no'] = self.env['ir.sequence'].next_by_code('export.logics')
 		vals['our_job_no'] = self.env['ir.sequence'].next_by_code('export.job.num')
-		print "Something"
 		new_record = super(Exportlogic, self).create(vals)
 
 		return new_record
@@ -163,12 +173,12 @@ class Exportlogic(models.Model):
 	@api.multi
 	def over(self):
 		self.state = "done"
-
-
+	
 	@api.multi
 	def create_sale(self):
 		prev_rec = self.env['sale.order'].search([('sales_id','=',self.id)])
-		prev_rec.unlink()
+		if prev_rec:
+			prev_rec.unlink()
 		# if not prev_rec:
 		value = 0 
 		get_id = self.env['product.template'].search([])
@@ -181,8 +191,10 @@ class Exportlogic(models.Model):
 		for data in self.export_id:
 			records = self.env['sale.order'].create({
 				'partner_id':self.customer.id,
+				'by_customer':self.by_customer.id,
 				'date_order':self.date,
 				'bill_type':self.bill_types,
+				'bill_no':self.bill_no,
 				'suppl_name':data.transporter.id,
 				'suppl_freight':data.trans_charge,
 				'form':data.form.name,
@@ -214,7 +226,7 @@ class Exportlogic(models.Model):
 
 		for line in lisst:
 			create_invoice = invoice.create({
-				'journal_id': 1,
+				'journal_id': 3,
 				'partner_id':line.id,
 				'customer':self.customer.id,
 				'date_invoice' : self.date,
@@ -243,12 +255,15 @@ class Exportlogic(models.Model):
 		if self.bill_types == "B/L Number":
 
 			create_invoice = invoice.create({
-				'journal_id': 1,
+				'journal_id': 3,
 				'partner_id':self.customer.id,
+				'by_customer':self.by_customer.id,
 				'date_invoice': self.date,
 				'billng_type':self.bill_types,
 				'bill_num':self.bill_no,
 				})
+
+			self.acc_link = create_invoice.id
 
 			for x in self.export_serv:
 				create_invoice_lines= invoice_lines.create({
@@ -267,12 +282,15 @@ class Exportlogic(models.Model):
 			
 
 			create_invoice = invoice.create({
-				'journal_id': 1,
+				'journal_id': 3,
 				'partner_id':self.customer.id,
+				'by_customer':self.by_customer.id,
 				'date_invoice': self.date,
 				'billng_type':self.bill_types,
 				'bill_num':self.bill_no,
 				})
+
+			self.acc_link = create_invoice.id
 
 			for line in data:
 				value = 0
@@ -322,7 +340,7 @@ class service_cont_tree(models.Model):
 
 	sevr_type_cont       = fields.Many2one('serv.types',string="Service Type")
 	sevr_charge_cont     = fields.Integer(string="Service Charges")
-	type_contt           = fields.Char(string="Container Type")
+	type_contt           = fields.Char(string="Container Size")
 
 	service_tree_cont    = fields.Many2one('export.logic')
 
@@ -340,7 +358,7 @@ class export_tree(models.Model):
 	custm_charge     = fields.Char(string="Customer Charges")
 	types            = fields.Selection([
 					('20 ft', '20 ft'),
-					('40 ft', '40 ft')],string="Type")
+					('40 ft', '40 ft')],string="Size")
 
 	crt_tree     = fields.Many2one('export.logic')
 
@@ -351,11 +369,11 @@ class export_tree(models.Model):
 		if self.transporter.id and self.form.id and self.to.id and self.fleet_type:
 			trans = self.env['res.partner'].search([('id','=',self.transporter.id)])
 			for x in trans.route_id:
-				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "exp":
+				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "export":
 					self.trans_charge = x.trans_charges
 			rec = self.env['res.partner'].search([('id','=',self.crt_tree.customer.id)])
 			for x in rec.route_id:
-				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "exp":
+				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "export":
 					self.custm_charge = x.trans_charges
 
 
@@ -402,6 +420,7 @@ class Importlogic(models.Model):
 
 
 	customer         = fields.Many2one('res.partner',string="Customer",required=True)
+	by_customer      = fields.Many2one('by.customer', string="By Customer")
 	bill_types       = fields.Char(string="Billing Type")
 	bill_bol         = fields.Boolean(string="B/L")
 	contt_bol        = fields.Boolean(string="B/L")
@@ -410,16 +429,21 @@ class Importlogic(models.Model):
 	job_no           = fields.Char(string="Job No", readonly=True)
 	date             = fields.Date(string="Date" ,required=True,default=date.today())
 	customer_ref     = fields.Char(string="Customer Ref")
+	cust_ref_inv     = fields.Char(string="Customer Ref Inv No")
 	site             = fields.Many2one('import.site',string="Site")
 	fri_id           = fields.Many2one('freight.forward', string="Freight Link")
 	shipper_date     = fields.Date(string="DOC Received Date",default=date.today())
 	vessel_date      = fields.Date(string="Vessel Arrival Date")
+	vessel_name      = fields.Char(string="Vessel Name")
+	s_supplier       = fields.Many2one('res.partner',string="Shipping Line")
 	bill_attach      = fields.Binary(string=" ")
 	bill_no          = fields.Char(string="B/L Number")
+	rot_no           = fields.Char(string="Rotation Number/Sequence Number")
 	twen_ft          = fields.Integer(string="20 ft")
 	fort_ft          = fields.Integer(string="40 ft")
 	do_attach        = fields.Binary(string=" ")
-	do_no            = fields.Char(string="Do No.")
+	do_no            = fields.Boolean(string="Do No.")
+	acc_link         = fields.Many2one('account.invoice',string="Invoice",readonly=True)
 	bayan_attach     = fields.Binary(string=" ")
 	final_bayan      = fields.Char(string="Final Bayan")
 	final_attach     = fields.Binary(string="Final Bayan")
@@ -431,6 +455,8 @@ class Importlogic(models.Model):
 	import_serv 	 = fields.One2many('import.service.tree','import_tree')
 	imp_contt 	     = fields.One2many('import.contain.tree','imp_tree_cont')
 	remarks          = fields.Text(string="Remarks")
+	eta              = fields.Date(string="ETA")
+	etd              = fields.Date(string="ETD")
 	stages 			 = fields.Selection([
 					 ('draft', 'Draft'),
 					 ('pre', 'Pre Bayan'),
@@ -439,6 +465,11 @@ class Importlogic(models.Model):
 					 ('done', 'Done'),
 					 ],default='draft')
 
+	_sql_constraints = [
+	('customer_ref', 'unique(customer_ref)','This customer reference already esixts!')
+	]
+
+	tick  = fields.Boolean()
 
 	@api.model
 	def create(self, vals):
@@ -471,46 +502,51 @@ class Importlogic(models.Model):
 			self.contt_bol = False
 
 
-	@api.onchange('customer','bill_types')
+	@api.onchange('customer','by_customer')
 	def get_import_tree_value(self):
 		if self.customer:
 			if self.bill_types == "B/L Number":
 				records = self.env['res.partner'].search([('id','=',self.customer.id)])
+				for x in records.bl_id:
+					if self.by_customer == x.by_customer:
 
-				import_serv = []
-				delete = []
-				delete = delete.append(2)
-				self.import_serv = delete
+						import_serv = []
+						delete = []
+						delete = delete.append(2)
+						self.import_serv = delete
 
-				inv = []
-				for invo in records.bl_id:
-					inv.append({
-						'charge_serv':invo.charges_serv,
-						'type_serv':invo.charges_type.id,
-						'import_tree':self.id,
-						})
-				
-				self.import_serv = inv
-				inv=[]
+						inv = []
+						for invo in x:
+							inv.append({
+								'charge_serv':invo.charges_serv,
+								'type_serv':invo.charges_type.id,
+								'import_tree':self.id,
+								})
+						
+						self.import_serv = inv
+						inv=[]
 
 			if self.bill_types == "Container Wise":
 				records = self.env['res.partner'].search([('id','=',self.customer.id)])
-				imp_contt = []
-				delete = []
-				delete = delete.append(2)
-				self.imp_contt = delete
+				for x in records.cont_id:
+					if self.by_customer == x.by_customer:
 
-				contt = []
-				for line in records.cont_id:
-					contt.append({
-						'sevr_charge_imp':line.charges_serv,
-						'sevr_type_imp':line.charges_type.id,
-						'type_contt_imp':line.cont_type,
-						'imp_tree_cont':self.id,
-						})
-				
-				self.imp_contt = contt
-				contt=[]
+						imp_contt = []
+						delete = []
+						delete = delete.append(2)
+						self.imp_contt = delete
+
+						contt = []
+						for line in x:
+							contt.append({
+								'sevr_charge_imp':line.charges_serv,
+								'sevr_type_imp':line.charges_type.id,
+								'type_contt_imp':line.cont_type,
+								'imp_tree_cont':self.id,
+								})
+						
+						self.imp_contt = contt
+						contt=[]
 
 
 	@api.multi
@@ -535,27 +571,29 @@ class Importlogic(models.Model):
 
 	@api.multi
 	def create_sale(self):
-		prev_rec = self.env['sale.order'].search([('sales_id','=',self.id)])
-		prev_rec.unlink()
-		# if not prev_rec:
+		prev_rec = self.env['sale.order'].search([('sales_imp_id','=',self.id)])
+		if prev_rec:
+			prev_rec.unlink()
 
 		value = 0 
 		get_id = self.env['product.template'].search([])
 		for x in get_id:
 			if x.name == "Container":
 				value = x.id
-				print value
+				
 
 		for data in self.import_id:
 			records = self.env['sale.order'].create({
 				'partner_id':self.customer.id,
+				'by_customer':self.by_customer.id,
 				'date_order':self.date,
 				'suppl_name':data.transporter.id,
 				'bill_type':self.bill_types,
+				'bill_no':self.bill_no,
 				'suppl_freight':data.trans_charge,
 				'form':data.form.name,
 				'to':data.to.name,
-				'sales_id': self.id,
+				'sales_imp_id': self.id,
 				})
 
 			records.order_line.create({
@@ -577,12 +615,15 @@ class Importlogic(models.Model):
 		if self.bill_types == "B/L Number":
 
 			create_invoice = invoice.create({
-				'journal_id': 1,
+				'journal_id': 3,
 				'partner_id':self.customer.id,
+				'by_customer':self.by_customer.id,
 				'date_invoice': self.date,
 				'billng_type':self.bill_types,
 				'bill_num':self.bill_no,
 				})
+
+			self.acc_link = create_invoice.id
 
 			for x in self.import_serv:
 				create_invoice_lines= invoice_lines.create({
@@ -601,12 +642,15 @@ class Importlogic(models.Model):
 			
 
 			create_invoice = invoice.create({
-				'journal_id': 1,
+				'journal_id': 3,
 				'partner_id':self.customer.id,
+				'by_customer':self.by_customer.id,
 				'date_invoice': self.date,
 				'billng_type':self.bill_types,
 				'bill_num':self.bill_no,
 				})
+
+			self.acc_link = create_invoice.id
 
 			for line in entry:
 				value = 0
@@ -645,7 +689,7 @@ class import_tree(models.Model):
 
 	types        	= fields.Selection([
 					('20 ft', '20 ft'),
-					('40 ft', '40 ft')],string="Type")
+					('40 ft', '40 ft')],string="Size")
 
 	crt_tree     = fields.Many2one('import.logic')
 
@@ -655,11 +699,11 @@ class import_tree(models.Model):
 		if self.transporter.id and self.form.id and self.to.id and self.fleet_type:
 			trans = self.env['res.partner'].search([('id','=',self.transporter.id)])
 			for x in trans.route_id:
-				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "imp":
+				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "import":
 					self.trans_charge = x.trans_charges
 			rec = self.env['res.partner'].search([('id','=',self.crt_tree.customer.id)])
 			for x in rec.route_id:
-				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "imp":
+				if self.form.id == x.form.id and self.to.id == x.to.id and self.fleet_type == x.fleet_type and x.service_type == "import":
 					self.custm_charge = x.trans_charges
 
 
@@ -677,7 +721,7 @@ class import_cont_tree(models.Model):
 
 	sevr_type_imp       = fields.Many2one('serv.types',string="Service Type")
 	sevr_charge_imp     = fields.Integer(string="Service Charges")
-	type_contt_imp      = fields.Char(string="Container Type")
+	type_contt_imp      = fields.Char(string="Container Size")
 
 	imp_tree_cont       = fields.Many2one('import.logic')
